+++ lib/app_root.dart
@@ -8,14 +8,14 @@ import 'package:storyzz/core/data/repository/auth_repository.dart';
 import 'package:storyzz/core/data/repository/maps_repository.dart';
 import 'package:storyzz/core/data/repository/setting_repository.dart';
 import 'package:storyzz/core/data/repository/story_repository.dart';
-import 'package:storyzz/core/provider/address_provider.dart';
-import 'package:storyzz/core/provider/auth_provider.dart';
-import 'package:storyzz/core/provider/settings_provider.dart';
-import 'package:storyzz/core/provider/story_provider.dart';
-import 'package:storyzz/core/provider/upload_story_provider.dart';
+import 'package:storyzz/core/providers/address_provider.dart';
+import 'package:storyzz/core/providers/auth_provider.dart';
+import 'package:storyzz/core/providers/settings_provider.dart';
+import 'package:storyzz/core/providers/story_provider.dart';
 import 'package:storyzz/core/routes/my_route_delegate.dart';
 import 'package:storyzz/core/routes/my_route_information_parser.dart';
 import 'package:storyzz/features/map/provider/map_provider.dart';
+import 'package:storyzz/features/upload_story/presentation/providers/upload_story_provider.dart';
 import 'package:storyzz/my_app.dart';
 
 class AppRoot extends StatelessWidget {
