+++ lib/core/providers/auth_provider.dart
