+++ lib/core/providers/settings_provider.dart
