+++ lib/core/providers/story_provider.dart
