+++ lib/features/map/provider/map_provider.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:storyzz/core/data/networking/responses/list_story.dart';
-import 'package:storyzz/core/provider/auth_provider.dart';
-import 'package:storyzz/core/provider/story_provider.dart';
+import 'package:storyzz/core/providers/auth_provider.dart';
+import 'package:storyzz/core/providers/story_provider.dart';
 import 'package:storyzz/features/map/services/map_service.dart';
 
 /// Manages map state, markers, and interactions for displaying user stories.
