+++ lib/features/settings/presentation/screen/settings_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:storyzz/core/localization/l10n/app_localizations.dart';
-import 'package:storyzz/core/provider/settings_provider.dart';
+import 'package:storyzz/core/providers/settings_provider.dart';
 import 'package:storyzz/core/widgets/language_selector.dart';
 import 'package:url_launcher/url_launcher.dart';
