+++ lib/features/upload_story/presentation/screen/upload_story_screen.dart
@@ -5,11 +5,11 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 import 'package:storyzz/core/localization/l10n/app_localizations.dart';
-import 'package:storyzz/core/provider/auth_provider.dart';
-import 'package:storyzz/core/provider/story_provider.dart';
-import 'package:storyzz/core/provider/upload_story_provider.dart';
+import 'package:storyzz/core/providers/auth_provider.dart';
+import 'package:storyzz/core/providers/story_provider.dart';
 import 'package:storyzz/core/routes/my_route_delegate.dart';
 import 'package:storyzz/core/variant/build_configuration.dart';
+import 'package:storyzz/features/upload_story/presentation/providers/upload_story_provider.dart';
 import 'package:storyzz/features/upload_story/presentation/widgets/camera_web_view.dart';
 import 'package:storyzz/features/upload_story/presentation/widgets/image_preview.dart';
 import 'package:storyzz/features/upload_story/presentation/widgets/location_map_selector.dart';
