+++ lib/features/upload_story/presentation/widgets/camera_web_view.dart
@@ -2,7 +2,7 @@ import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:storyzz/core/localization/l10n/app_localizations.dart';
-import 'package:storyzz/core/provider/upload_story_provider.dart';
+import 'package:storyzz/features/upload_story/presentation/providers/upload_story_provider.dart';
 import 'package:storyzz/features/upload_story/services/camera_service.dart';
 
 class CameraViewWeb extends StatelessWidget {
