+++ lib/features/upload_story/services/image_picker_services.dart
@@ -1,10 +1,11 @@
 import 'dart:io';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 import 'package:storyzz/core/localization/l10n/app_localizations.dart';
-import 'package:storyzz/core/provider/upload_story_provider.dart';
+import 'package:storyzz/features/upload_story/presentation/providers/upload_story_provider.dart';
 
 class ImagePickerService {
   final ImagePicker _picker = ImagePicker();
